Fix linting Issue
 - Separate the button logic components/ui/button.tsx to components/ui/Button folder
 - components/ui/Button/index.tsx => Button Component
 - components/ui/Button/variants.tsx => Button Variants
